+++ src/__tests__/prepublish.test.js
@@ -20,7 +20,7 @@ describe('PREPUBLISH command', () => {
 
   it('copies READMEs as appropriate', async () => {
     const helpers = require('../utils/shell');
-    await prepublish({ src: 'test/fixtures/packages/*' });
+    await prepublish({ src: 'test/fixtures/packages/*', copyAttrs: COPY_SPECS.join(',') });
     const normalizedArgs = helpers.cp.mock.calls.map(([src, dst]) =>
       [normalizePath(src), normalizePath(dst)]
     );
@@ -30,16 +30,27 @@ describe('PREPUBLISH command', () => {
   it('copies common attributes to subpackages', async () => {
     const writeSpecs = require('../utils/writeSpecs').default;
     const refSpecs = require(path.join(process.cwd(), 'package.json'));
-    await prepublish({ src: 'test/fixtures/packages/*' });
+    await prepublish({ src: 'test/fixtures/packages/*', copyAttrs: COPY_SPECS.join(',') });
     expect(writeSpecs.mock.calls.map((args) => args[1].name)).toEqual(['oao', 'oao-b', 'oao-c', 'oao-d']);
     writeSpecs.mock.calls.forEach(([, specs]) => {
       COPY_SPECS.forEach((attr) => { expect(specs[attr]).toEqual(refSpecs[attr]); });
     });
   });
 
+  it('allows configuring attributes to be copied to subpackages', async () => {
+    const writeSpecs = require('../utils/writeSpecs').default;
+    const refSpecs = require(path.join(process.cwd(), 'package.json'));
+    await prepublish({ src: 'test/fixtures/packages/*', copyAttrs: 'author' });
+    expect(writeSpecs.mock.calls.map((args) => args[1].name)).toEqual(['oao', 'oao-b', 'oao-c', 'oao-d']);
+    writeSpecs.mock.calls.forEach(([, specs]) => {
+      expect(specs.author).toEqual(refSpecs.author);
+      expect(specs.description).toBeUndefined();
+    });
+  });
+
   it('throws when a package has a version > master', async () => {
     try {
-      await prepublish({ src: 'test/fixtures/packagesWrongVersion/*' });
+      await prepublish({ src: 'test/fixtures/packagesWrongVersion/*', copyAttrs: COPY_SPECS.join(',') });
       throw new Error('DID_NOT_THROW');
     } catch (err) {
       if (err.message === 'DID_NOT_THROW') throw err;
@@ -48,7 +59,7 @@ describe('PREPUBLISH command', () => {
 
   it('throws when a package has an invalid version', async () => {
     try {
-      await prepublish({ src: 'test/fixtures/packagesWrongVersion2/*' });
+      await prepublish({ src: 'test/fixtures/packagesWrongVersion2/*', copyAttrs: COPY_SPECS.join(',') });
       throw new Error('DID_NOT_THROW');
     } catch (err) {
       if (err.message === 'DID_NOT_THROW') throw err;